+++ src/main/java/com/slowflow/slowflowbackend/SlowFlowBackendApplication.java
@@ -1,4 +1,4 @@
-package com.slowflow.SlowFlow_backend;
+package com.slowflow.slowflowbackend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
