+++ src/main/java/com/slowflow/slowflowbackend/score/model/DailyState.java
@@ -1,4 +1,4 @@
-package com.slowflow.SlowFlow_backend.score.model;
+package com.slowflow.slowflowbackend.score.model;
 
 public enum DailyState {
     NONE, NEGATIVE, NEUTRAL, POSITIVE, GOAL_REACHED
