+++ src/test/java/com/slowflow/slowflowbackend/SlowFlowBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.slowflow.SlowFlow_backend;
+package com.slowflow.slowflowbackend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
